Fixed compiler errors and warnings
Changes:
   - Added numbers header to Constants.h
   - Removed pragma region in the SwerveModule.hpp file
   - Disabled desktop support to remove RevLib linker errors

+++ src/main/cpp/PortManager.cpp
@@ -18,12 +18,12 @@ PortManager& PortManager::Instance()
     return instance;
 }
 
-int PortManager::GetCanID(std::string IDName)
+unsigned int PortManager::GetCanID(std::string IDName)
 {
     int ID = m_CANPorts[IDName];
 
     if (ID < 0)
         FRC_ReportError(frc::warn::Warning, "Requested CAN ID was {}. CAN IDs can not be less than 0.", ID);
 
-    return ID;
+    return (unsigned int)ID;
 }
